Split list expression plan modules
- Move typed list expression, item, element, local, and case helpers into role-owned child modules.
- Keep the ListExpr facade in the parent module and preserve existing plan behavior.
- Rehome list expression unit tests beside the modules that own the reviewed shapes.
