+++ src/plan/expression/list.rs
@@ -1,17 +1,30 @@
+mod case;
+mod elements;
+mod item;
+mod local;
+mod typed;
+
+use self::case::{
+    into_bool_clauses, into_float_clauses, into_function_clauses, into_int_clauses,
+    into_list_clauses, into_nil_clauses, into_string_clauses, into_tuple_clauses,
+};
+use self::item::ListItemTag;
+pub(crate) use self::{
+    case::BoolListCaseBranches,
+    elements::{ListElementTypeMismatch, ListElements},
+    item::{
+        BoolListItem, FloatListItem, FunctionListItem, IntListItem, ListItem, ListListItem,
+        NilListItem, StringListItem, TupleListItem,
+    },
+    local::ListLocalExpr,
+    typed::{TypedListExpr, TypedListExprKind},
+};
 use super::{
-    BoolExpr, CallArg, Expr, ExprKind, FloatExpr, FunctionExpr, IntExpr, ListFunctionExpr, NilExpr,
-    PanicExpr, StringExpr, TupleExpr,
+    BoolExpr, CallArg, Expr, FloatExpr, IntExpr, ListFunctionExpr, PanicExpr, StringExpr, TupleExpr,
 };
-use crate::plan::{
-    BoolListFunctionId, BoolListLocalId, FloatListFunctionId, FloatListLocalId,
-    FunctionListFunctionId, FunctionListLocalId, FunctionType, IntListFunctionId, IntListLocalId,
-    ListFunctionId, ListListFunctionId, ListListLocalId, ListLocal, NilListFunctionId,
-    NilListLocalId, Step, StringListFunctionId, StringListLocalId, TupleListFunctionId,
-    TupleListLocalId, ValueType,
-};
+use crate::plan::{ListFunctionId, ListLocal, Step, ValueType};
 use ecow::EcoString;
 use num_bigint::BigInt;
-use std::fmt::Debug;
 
 #[derive(Debug, Clone, PartialEq)]
 pub(crate) enum ListExpr {
@@ -33,411 +46,6 @@ pub(crate) type NilListExpr = TypedListExpr<NilListItem>;
 pub(crate) type TupleListExpr = TypedListExpr<TupleListItem>;
 pub(crate) type ListListExpr = TypedListExpr<ListListItem>;
 pub(crate) type FunctionListExpr = TypedListExpr<FunctionListItem>;
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum BoolListCaseBranches {
-    Int {
-        true_: IntListExpr,
-        false_: IntListExpr,
-    },
-    String {
-        true_: StringListExpr,
-        false_: StringListExpr,
-    },
-    Float {
-        true_: FloatListExpr,
-        false_: FloatListExpr,
-    },
-    Bool {
-        true_: BoolListExpr,
-        false_: BoolListExpr,
-    },
-    Nil {
-        true_: NilListExpr,
-        false_: NilListExpr,
-    },
-    Tuple {
-        true_: TupleListExpr,
-        false_: TupleListExpr,
-    },
-    List {
-        true_: ListListExpr,
-        false_: ListListExpr,
-    },
-    Function {
-        true_: FunctionListExpr,
-        false_: FunctionListExpr,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct TypedListExpr<Item: ListItem> {
-    item: Item,
-    kind: TypedListExprKind<Item>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum TypedListExprKind<Item: ListItem> {
-    Value(Vec<Item::ElementExpr>),
-    Spread {
-        elements: Vec<Item::ElementExpr>,
-        tail: Box<TypedListExpr<Item>>,
-    },
-    LocalGet {
-        local: Item::Local,
-        name: EcoString,
-    },
-    Call {
-        function: Item::Function,
-        args: Vec<CallArg>,
-    },
-    FunctionCall {
-        function: Box<ListFunctionExpr>,
-        args: Vec<CallArg>,
-    },
-    TupleIndex {
-        tuple: Box<TupleExpr>,
-        index: usize,
-    },
-    ListIndex {
-        list: Box<ListListExpr>,
-        index: usize,
-    },
-    DropFirst {
-        list: Box<TypedListExpr<Item>>,
-        count: usize,
-    },
-    Panic(PanicExpr),
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<TypedListExpr<Item>>,
-        false_: Box<TypedListExpr<Item>>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, TypedListExpr<Item>)>,
-        fallback: Box<TypedListExpr<Item>>,
-    },
-    StringCase {
-        subject: Box<StringExpr>,
-        clauses: Vec<(EcoString, TypedListExpr<Item>)>,
-        fallback: Box<TypedListExpr<Item>>,
-    },
-    FloatCase {
-        subject: Box<FloatExpr>,
-        clauses: Vec<(f64, TypedListExpr<Item>)>,
-        fallback: Box<TypedListExpr<Item>>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<TypedListExpr<Item>>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum ListElements {
-    Int(Vec<IntExpr>),
-    String(Vec<StringExpr>),
-    Float(Vec<FloatExpr>),
-    Bool(Vec<BoolExpr>),
-    Nil(Vec<NilExpr>),
-    Tuple {
-        item_type: Vec<ValueType>,
-        values: Vec<TupleExpr>,
-    },
-    List {
-        item_type: Box<ValueType>,
-        values: Vec<ListExpr>,
-    },
-    Function {
-        item_type: FunctionType,
-        values: Vec<FunctionExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum ListLocalExpr {
-    Int {
-        local: IntListLocalId,
-        value: IntListExpr,
-    },
-    String {
-        local: StringListLocalId,
-        value: StringListExpr,
-    },
-    Float {
-        local: FloatListLocalId,
-        value: FloatListExpr,
-    },
-    Bool {
-        local: BoolListLocalId,
-        value: BoolListExpr,
-    },
-    Nil {
-        local: NilListLocalId,
-        value: NilListExpr,
-    },
-    Tuple {
-        local: TupleListLocalId,
-        item_type: Vec<ValueType>,
-        value: TupleListExpr,
-    },
-    List {
-        local: ListListLocalId,
-        item_type: Box<ValueType>,
-        value: ListListExpr,
-    },
-    Function {
-        local: FunctionListLocalId,
-        item_type: FunctionType,
-        value: FunctionListExpr,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct ListElementTypeMismatch {
-    pub(crate) expected: ValueType,
-    pub(crate) actual: ValueType,
-}
-
-pub(crate) trait ListItem: Debug + Clone + PartialEq {
-    type ElementExpr: Debug + Clone + PartialEq;
-    type Local: Debug + Clone + PartialEq;
-    type Function: Debug + Clone + PartialEq;
-
-    fn value_type(&self) -> ValueType;
-
-    fn local_to_facade(&self, local: Self::Local) -> ListLocal;
-
-    fn elements_from_exprs(
-        item: &Self,
-        values: Vec<Expr>,
-    ) -> Result<Vec<Self::ElementExpr>, ListElementTypeMismatch>;
-
-    fn elements_to_facade(item: Self, values: Vec<Self::ElementExpr>) -> ListElements;
-
-    fn expr_to_facade(expression: TypedListExpr<Self>) -> ListExpr
-    where
-        Self: Sized;
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct IntListItem;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct StringListItem;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct FloatListItem;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct BoolListItem;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct NilListItem;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct TupleListItem {
-    item_type: Vec<ValueType>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct ListListItem {
-    item_type: Box<ValueType>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct FunctionListItem {
-    item_type: FunctionType,
-}
-
-impl TupleListItem {
-    pub(crate) fn new(item_type: Vec<ValueType>) -> Self {
-        Self { item_type }
-    }
-
-    pub(crate) fn item_type(&self) -> Vec<ValueType> {
-        self.item_type.clone()
-    }
-}
-
-impl ListListItem {
-    pub(crate) fn new(item_type: Box<ValueType>) -> Self {
-        Self { item_type }
-    }
-
-    pub(crate) fn item_type(&self) -> Box<ValueType> {
-        self.item_type.clone()
-    }
-}
-
-impl FunctionListItem {
-    pub(crate) fn new(item_type: FunctionType) -> Self {
-        Self { item_type }
-    }
-
-    pub(crate) fn item_type(&self) -> FunctionType {
-        self.item_type.clone()
-    }
-}
-
-impl<Item: ListItem> TypedListExpr<Item> {
-    fn new(item: Item, kind: TypedListExprKind<Item>) -> Self {
-        Self { item, kind }
-    }
-
-    pub(crate) fn item(&self) -> &Item {
-        &self.item
-    }
-
-    pub(crate) fn element_type(&self) -> ValueType {
-        self.item.value_type()
-    }
-
-    pub(crate) fn kind(&self) -> &TypedListExprKind<Item> {
-        &self.kind
-    }
-
-    fn value(item: Item, elements: Vec<Item::ElementExpr>) -> Self {
-        Self::new(item, TypedListExprKind::Value(elements))
-    }
-
-    fn spread(item: Item, elements: Vec<Item::ElementExpr>, tail: TypedListExpr<Item>) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::Spread {
-                elements,
-                tail: Box::new(tail),
-            },
-        )
-    }
-
-    pub(crate) fn local_get(item: Item, local: Item::Local, name: EcoString) -> Self {
-        Self::new(item, TypedListExprKind::LocalGet { local, name })
-    }
-
-    fn call(item: Item, function: Item::Function, args: Vec<CallArg>) -> Self {
-        Self::new(item, TypedListExprKind::Call { function, args })
-    }
-
-    fn function_call(item: Item, function: ListFunctionExpr, args: Vec<CallArg>) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::FunctionCall {
-                function: Box::new(function),
-                args,
-            },
-        )
-    }
-
-    fn tuple_index(item: Item, tuple: TupleExpr, index: usize) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::TupleIndex {
-                tuple: Box::new(tuple),
-                index,
-            },
-        )
-    }
-
-    fn list_index(item: Item, list: ListListExpr, index: usize) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::ListIndex {
-                list: Box::new(list),
-                index,
-            },
-        )
-    }
-
-    fn drop_first(item: Item, list: TypedListExpr<Item>, count: usize) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::DropFirst {
-                list: Box::new(list),
-                count,
-            },
-        )
-    }
-
-    fn panic(item: Item, panic: PanicExpr) -> Self {
-        Self::new(item, TypedListExprKind::Panic(panic))
-    }
-
-    fn bool_case(
-        item: Item,
-        subject: BoolExpr,
-        true_: TypedListExpr<Item>,
-        false_: TypedListExpr<Item>,
-    ) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        )
-    }
-
-    fn int_case(
-        item: Item,
-        subject: IntExpr,
-        clauses: Vec<(BigInt, TypedListExpr<Item>)>,
-        fallback: TypedListExpr<Item>,
-    ) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        )
-    }
-
-    fn string_case(
-        item: Item,
-        subject: StringExpr,
-        clauses: Vec<(EcoString, TypedListExpr<Item>)>,
-        fallback: TypedListExpr<Item>,
-    ) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::StringCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        )
-    }
-
-    fn float_case(
-        item: Item,
-        subject: FloatExpr,
-        clauses: Vec<(f64, TypedListExpr<Item>)>,
-        fallback: TypedListExpr<Item>,
-    ) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::FloatCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        )
-    }
-
-    fn block(item: Item, steps: Vec<Step>, return_: TypedListExpr<Item>) -> Self {
-        Self::new(
-            item,
-            TypedListExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        )
-    }
-}
 
 impl ListExpr {
     pub(crate) fn value(elements: Vec<Expr>, element_type: ValueType) -> Self {
@@ -1080,515 +688,12 @@ impl ListExpr {
 }
 
 #[cfg(test)]
-macro_rules! impl_typed_list_expr_from_facade {
-    ($type:ty, $method:ident, $name:literal) => {
-        impl From<ListExpr> for $type {
-            fn from(value: ListExpr) -> Self {
-                value
-                    .$method()
-                    .expect(concat!("expected ", $name, " list expression"))
-            }
-        }
-    };
-}
-
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(IntListExpr, into_int, "int");
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(StringListExpr, into_string, "string");
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(FloatListExpr, into_float, "float");
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(BoolListExpr, into_bool, "bool");
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(NilListExpr, into_nil, "nil");
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(TupleListExpr, into_tuple, "tuple");
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(ListListExpr, into_list, "list");
-#[cfg(test)]
-impl_typed_list_expr_from_facade!(FunctionListExpr, into_function, "function");
-
-impl ListElements {
-    pub(crate) fn from_exprs(
-        item_type: ValueType,
-        values: Vec<Expr>,
-    ) -> Result<Self, ListElementTypeMismatch> {
-        match item_type {
-            ValueType::Int => list_elements_from_exprs(IntListItem, values),
-            ValueType::String => list_elements_from_exprs(StringListItem, values),
-            ValueType::Float => list_elements_from_exprs(FloatListItem, values),
-            ValueType::Bool => list_elements_from_exprs(BoolListItem, values),
-            ValueType::Nil => list_elements_from_exprs(NilListItem, values),
-            ValueType::Tuple(item_type) => {
-                list_elements_from_exprs(TupleListItem { item_type }, values)
-            }
-            ValueType::List(item_type) => {
-                list_elements_from_exprs(ListListItem { item_type }, values)
-            }
-            ValueType::Function(item_type) => list_elements_from_exprs(
-                FunctionListItem {
-                    item_type: *item_type,
-                },
-                values,
-            ),
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn item_type(&self) -> ValueType {
-        match self {
-            Self::Int(_) => ValueType::Int,
-            Self::String(_) => ValueType::String,
-            Self::Float(_) => ValueType::Float,
-            Self::Bool(_) => ValueType::Bool,
-            Self::Nil(_) => ValueType::Nil,
-            Self::Tuple { item_type, .. } => ValueType::Tuple(item_type.clone()),
-            Self::List { item_type, .. } => ValueType::List(item_type.clone()),
-            Self::Function { item_type, .. } => ValueType::Function(Box::new(item_type.clone())),
-        }
-    }
-}
-
-fn list_elements_from_exprs<Item: ListItem>(
-    item: Item,
-    values: Vec<Expr>,
-) -> Result<ListElements, ListElementTypeMismatch> {
-    let elements = Item::elements_from_exprs(&item, values)?;
-    Ok(Item::elements_to_facade(item, elements))
-}
-
-#[derive(Debug, Clone, PartialEq)]
-enum ListItemTag {
-    Int(IntListItem),
-    String(StringListItem),
-    Float(FloatListItem),
-    Bool(BoolListItem),
-    Nil(NilListItem),
-    Tuple(TupleListItem),
-    List(ListListItem),
-    Function(FunctionListItem),
-}
-
-impl ListItemTag {
-    fn from_value_type(value: ValueType) -> Self {
-        match value {
-            ValueType::Int => Self::Int(IntListItem),
-            ValueType::String => Self::String(StringListItem),
-            ValueType::Float => Self::Float(FloatListItem),
-            ValueType::Bool => Self::Bool(BoolListItem),
-            ValueType::Nil => Self::Nil(NilListItem),
-            ValueType::Tuple(item_type) => Self::Tuple(TupleListItem { item_type }),
-            ValueType::List(item_type) => Self::List(ListListItem { item_type }),
-            ValueType::Function(item_type) => Self::Function(FunctionListItem {
-                item_type: *item_type,
-            }),
-        }
-    }
-
-    fn expr_from_values(self, values: Vec<Expr>) -> Result<ListExpr, ListElementTypeMismatch> {
-        match self {
-            Self::Int(item) => ItemExprBuilder(item).build(values).map(ListExpr::Int),
-            Self::String(item) => ItemExprBuilder(item).build(values).map(ListExpr::String),
-            Self::Float(item) => ItemExprBuilder(item).build(values).map(ListExpr::Float),
-            Self::Bool(item) => ItemExprBuilder(item).build(values).map(ListExpr::Bool),
-            Self::Nil(item) => ItemExprBuilder(item).build(values).map(ListExpr::Nil),
-            Self::Tuple(item) => ItemExprBuilder(item).build(values).map(ListExpr::Tuple),
-            Self::List(item) => ItemExprBuilder(item).build(values).map(ListExpr::List),
-            Self::Function(item) => ItemExprBuilder(item).build(values).map(ListExpr::Function),
-        }
-    }
-}
-
-struct ItemExprBuilder<Item>(Item);
-
-impl<Item: ListItem> ItemExprBuilder<Item> {
-    fn build(self, values: Vec<Expr>) -> Result<TypedListExpr<Item>, ListElementTypeMismatch> {
-        let elements = Item::elements_from_exprs(&self.0, values)?;
-        Ok(TypedListExpr::value(self.0, elements))
-    }
-}
-
-macro_rules! primitive_list_item {
-    (
-        $item:ident,
-        $expr:ty,
-        $local:ty,
-        $function:ty,
-        $value_type:expr,
-        $expr_pattern:pat => $expr_value:expr,
-        $elements_variant:ident,
-        $facade_variant:path,
-        $local_ctor:path
-    ) => {
-        impl ListItem for $item {
-            type ElementExpr = $expr;
-            type Local = $local;
-            type Function = $function;
-
-            fn value_type(&self) -> ValueType {
-                $value_type
-            }
-
-            fn local_to_facade(&self, local: Self::Local) -> ListLocal {
-                $local_ctor(local)
-            }
-
-            fn elements_from_exprs(
-                _item: &Self,
-                values: Vec<Expr>,
-            ) -> Result<Vec<Self::ElementExpr>, ListElementTypeMismatch> {
-                values
-                    .into_iter()
-                    .map(|value| match value {
-                        Expr {
-                            kind: $expr_pattern,
-                        } => Ok($expr_value),
-                        value => Err(ListElementTypeMismatch {
-                            expected: $value_type,
-                            actual: value.value_type(),
-                        }),
-                    })
-                    .collect()
-            }
-
-            fn elements_to_facade(_item: Self, values: Vec<Self::ElementExpr>) -> ListElements {
-                ListElements::$elements_variant(values)
-            }
-
-            fn expr_to_facade(expression: TypedListExpr<Self>) -> ListExpr {
-                $facade_variant(expression)
-            }
-        }
-    };
-}
-
-primitive_list_item!(
-    IntListItem,
-    IntExpr,
-    IntListLocalId,
-    IntListFunctionId,
-    ValueType::Int,
-    ExprKind::Int(value) => value,
-    Int,
-    ListExpr::Int,
-    ListLocal::Int
-);
-
-primitive_list_item!(
-    StringListItem,
-    StringExpr,
-    StringListLocalId,
-    StringListFunctionId,
-    ValueType::String,
-    ExprKind::String(value) => value,
-    String,
-    ListExpr::String,
-    ListLocal::String
-);
-
-primitive_list_item!(
-    FloatListItem,
-    FloatExpr,
-    FloatListLocalId,
-    FloatListFunctionId,
-    ValueType::Float,
-    ExprKind::Float(value) => value,
-    Float,
-    ListExpr::Float,
-    ListLocal::Float
-);
-
-primitive_list_item!(
-    BoolListItem,
-    BoolExpr,
-    BoolListLocalId,
-    BoolListFunctionId,
-    ValueType::Bool,
-    ExprKind::Bool(value) => value,
-    Bool,
-    ListExpr::Bool,
-    ListLocal::Bool
-);
-
-primitive_list_item!(
-    NilListItem,
-    NilExpr,
-    NilListLocalId,
-    NilListFunctionId,
-    ValueType::Nil,
-    ExprKind::Nil(value) => value,
-    Nil,
-    ListExpr::Nil,
-    ListLocal::Nil
-);
-
-impl ListItem for TupleListItem {
-    type ElementExpr = TupleExpr;
-    type Local = TupleListLocalId;
-    type Function = TupleListFunctionId;
-
-    fn value_type(&self) -> ValueType {
-        ValueType::Tuple(self.item_type.clone())
-    }
-
-    fn local_to_facade(&self, local: Self::Local) -> ListLocal {
-        ListLocal::tuple(local, self.item_type.clone())
-    }
-
-    fn elements_from_exprs(
-        item: &Self,
-        values: Vec<Expr>,
-    ) -> Result<Vec<Self::ElementExpr>, ListElementTypeMismatch> {
-        values
-            .into_iter()
-            .map(|value| match value {
-                Expr {
-                    kind: ExprKind::Tuple(value),
-                } if value.type_() == item.item_type.as_slice() => Ok(value),
-                value => Err(ListElementTypeMismatch {
-                    expected: item.value_type(),
-                    actual: value.value_type(),
-                }),
-            })
-            .collect()
-    }
-
-    fn elements_to_facade(item: Self, values: Vec<Self::ElementExpr>) -> ListElements {
-        ListElements::Tuple {
-            item_type: item.item_type,
-            values,
-        }
-    }
-
-    fn expr_to_facade(expression: TypedListExpr<Self>) -> ListExpr {
-        ListExpr::Tuple(expression)
-    }
-}
-
-impl ListItem for ListListItem {
-    type ElementExpr = ListExpr;
-    type Local = ListListLocalId;
-    type Function = ListListFunctionId;
-
-    fn value_type(&self) -> ValueType {
-        ValueType::List(self.item_type.clone())
-    }
-
-    fn local_to_facade(&self, local: Self::Local) -> ListLocal {
-        ListLocal::list(local, self.item_type.as_ref().clone())
-    }
-
-    fn elements_from_exprs(
-        item: &Self,
-        values: Vec<Expr>,
-    ) -> Result<Vec<Self::ElementExpr>, ListElementTypeMismatch> {
-        values
-            .into_iter()
-            .map(|value| match value {
-                Expr {
-                    kind: ExprKind::List(value),
-                } if value.element_type() == item.item_type.as_ref().clone() => Ok(value),
-                value => Err(ListElementTypeMismatch {
-                    expected: item.value_type(),
-                    actual: value.value_type(),
-                }),
-            })
-            .collect()
-    }
-
-    fn elements_to_facade(item: Self, values: Vec<Self::ElementExpr>) -> ListElements {
-        ListElements::List {
-            item_type: item.item_type,
-            values,
-        }
-    }
-
-    fn expr_to_facade(expression: TypedListExpr<Self>) -> ListExpr {
-        ListExpr::List(expression)
-    }
-}
-
-impl ListItem for FunctionListItem {
-    type ElementExpr = FunctionExpr;
-    type Local = FunctionListLocalId;
-    type Function = FunctionListFunctionId;
-
-    fn value_type(&self) -> ValueType {
-        ValueType::Function(Box::new(self.item_type.clone()))
-    }
-
-    fn local_to_facade(&self, local: Self::Local) -> ListLocal {
-        ListLocal::function(local, self.item_type.clone())
-    }
-
-    fn elements_from_exprs(
-        item: &Self,
-        values: Vec<Expr>,
-    ) -> Result<Vec<Self::ElementExpr>, ListElementTypeMismatch> {
-        values
-            .into_iter()
-            .map(|value| match value {
-                Expr {
-                    kind: ExprKind::Function(value),
-                } if value.type_() == &item.item_type => Ok(value),
-                value => Err(ListElementTypeMismatch {
-                    expected: item.value_type(),
-                    actual: value.value_type(),
-                }),
-            })
-            .collect()
-    }
-
-    fn elements_to_facade(item: Self, values: Vec<Self::ElementExpr>) -> ListElements {
-        ListElements::Function {
-            item_type: item.item_type,
-            values,
-        }
-    }
-
-    fn expr_to_facade(expression: TypedListExpr<Self>) -> ListExpr {
-        ListExpr::Function(expression)
-    }
-}
-
-fn into_int_clauses<Pattern>(clauses: Vec<(Pattern, ListExpr)>) -> Vec<(Pattern, IntListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            (
-                pattern,
-                branch
-                    .into_int()
-                    .expect("list case branches must be List(Int)"),
-            )
-        })
-        .collect()
-}
-
-fn into_string_clauses<Pattern>(
-    clauses: Vec<(Pattern, ListExpr)>,
-) -> Vec<(Pattern, StringListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            (
-                pattern,
-                branch
-                    .into_string()
-                    .expect("list case branches must be List(String)"),
-            )
-        })
-        .collect()
-}
-
-fn into_float_clauses<Pattern>(clauses: Vec<(Pattern, ListExpr)>) -> Vec<(Pattern, FloatListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            (
-                pattern,
-                branch
-                    .into_float()
-                    .expect("list case branches must be List(Float)"),
-            )
-        })
-        .collect()
-}
-
-fn into_bool_clauses<Pattern>(clauses: Vec<(Pattern, ListExpr)>) -> Vec<(Pattern, BoolListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            (
-                pattern,
-                branch
-                    .into_bool()
-                    .expect("list case branches must be List(Bool)"),
-            )
-        })
-        .collect()
-}
-
-fn into_nil_clauses<Pattern>(clauses: Vec<(Pattern, ListExpr)>) -> Vec<(Pattern, NilListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            (
-                pattern,
-                branch
-                    .into_nil()
-                    .expect("list case branches must be List(Nil)"),
-            )
-        })
-        .collect()
-}
-
-fn into_tuple_clauses<Pattern>(
-    clauses: Vec<(Pattern, ListExpr)>,
-    item: &TupleListItem,
-) -> Vec<(Pattern, TupleListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            let branch = branch
-                .into_tuple()
-                .expect("list case branches must be List(tuple)");
-            assert_eq!(
-                &branch.item, item,
-                "tuple list branch item types must match"
-            );
-            (pattern, branch)
-        })
-        .collect()
-}
-
-fn into_list_clauses<Pattern>(
-    clauses: Vec<(Pattern, ListExpr)>,
-    item: &ListListItem,
-) -> Vec<(Pattern, ListListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            let branch = branch
-                .into_list()
-                .expect("list case branches must be List(list)");
-            assert_eq!(
-                &branch.item, item,
-                "nested list branch item types must match"
-            );
-            (pattern, branch)
-        })
-        .collect()
-}
-
-fn into_function_clauses<Pattern>(
-    clauses: Vec<(Pattern, ListExpr)>,
-    item: &FunctionListItem,
-) -> Vec<(Pattern, FunctionListExpr)> {
-    clauses
-        .into_iter()
-        .map(|(pattern, branch)| {
-            let branch = branch
-                .into_function()
-                .expect("list case branches must be List(function)");
-            assert_eq!(
-                &branch.item, item,
-                "function list branch item types must match"
-            );
-            (pattern, branch)
-        })
-        .collect()
-}
-
-#[cfg(test)]
 mod tests {
     use super::{
         BoolListCaseBranches, BoolListExpr, BoolListItem, FloatListExpr, FloatListItem,
         FunctionListExpr, FunctionListItem, IntListExpr, IntListItem, ListElementTypeMismatch,
         ListElements, ListExpr, ListListExpr, ListListItem, NilListExpr, NilListItem,
-        StringListExpr, StringListItem, TupleListExpr, TupleListItem, TypedListExprKind,
-        into_function_clauses, into_list_clauses, into_nil_clauses, into_tuple_clauses,
+        StringListExpr, StringListItem, TupleListExpr, TupleListItem,
     };
     use crate::plan::{
         BoolExpr, Expr, FloatExpr, FunctionExpr, FunctionType, FunctionValue, IntExpr,
@@ -1687,71 +792,6 @@ mod tests {
     }
 
     #[test]
-    fn from_exprs_rejects_wrong_item_family_and_nested_metadata() {
-        assert_eq!(
-            ListElements::from_exprs(
-                ValueType::Int,
-                vec![Expr::string(StringExpr::value("wrong".into()))]
-            ),
-            Err(ListElementTypeMismatch {
-                expected: ValueType::Int,
-                actual: ValueType::String,
-            }),
-        );
-
-        let tuple = TupleExpr::value(
-            vec![Expr::string(StringExpr::value("wrong".into()))],
-            vec![ValueType::String],
-        );
-        assert_eq!(
-            ListElements::from_exprs(
-                ValueType::Tuple(vec![ValueType::Int]),
-                vec![Expr::tuple(tuple)],
-            ),
-            Err(ListElementTypeMismatch {
-                expected: ValueType::Tuple(vec![ValueType::Int]),
-                actual: ValueType::Tuple(vec![ValueType::String]),
-            }),
-        );
-
-        let nested = ListExpr::value(
-            vec![Expr::string(StringExpr::value("wrong".into()))],
-            ValueType::String,
-        );
-        assert_eq!(
-            ListElements::from_exprs(
-                ValueType::List(Box::new(ValueType::Int)),
-                vec![Expr::list(nested)],
-            ),
-            Err(ListElementTypeMismatch {
-                expected: ValueType::List(Box::new(ValueType::Int)),
-                actual: ValueType::List(Box::new(ValueType::String)),
-            }),
-        );
-
-        let function = FunctionExpr::value(FunctionValue::new(
-            RuntimeFunctionId::Int(IntFunctionId(0)),
-            Vec::new(),
-        ));
-        assert_eq!(
-            ListElements::from_exprs(
-                ValueType::Function(Box::new(FunctionType::new(Vec::new(), ValueType::String))),
-                vec![Expr::function(function)],
-            ),
-            Err(ListElementTypeMismatch {
-                expected: ValueType::Function(Box::new(FunctionType::new(
-                    Vec::new(),
-                    ValueType::String,
-                ))),
-                actual: ValueType::Function(Box::new(FunctionType::new(
-                    Vec::new(),
-                    ValueType::Int,
-                ))),
-            }),
-        );
-    }
-
-    #[test]
     fn spread_constructor_preserves_typed_tail_family() {
         let int_tail = ListExpr::value(vec![Expr::int(IntExpr::value(2.into()))], ValueType::Int);
         assert_eq!(
@@ -1809,6 +849,27 @@ mod tests {
                 actual: ValueType::String,
             }),
         );
+    }
+
+    #[test]
+    fn facade_typed_projection_rejects_wrong_item_family() {
+        assert_eq!(
+            ListExpr::value(
+                vec![Expr::string(StringExpr::value("one".into()))],
+                ValueType::String,
+            )
+            .into_int(),
+            None,
+        );
+
+        let int_list = ListExpr::value(vec![Expr::int(IntExpr::value(1.into()))], ValueType::Int);
+        assert_eq!(int_list.clone().into_string(), None);
+        assert_eq!(int_list.clone().into_float(), None);
+        assert_eq!(int_list.clone().into_bool(), None);
+        assert_eq!(int_list.clone().into_nil(), None);
+        assert_eq!(int_list.clone().into_tuple(), None);
+        assert_eq!(int_list.clone().into_list(), None);
+        assert_eq!(int_list.into_function(), None);
     }
 
     #[test]
@@ -1943,67 +1004,6 @@ mod tests {
     }
 
     #[test]
-    fn typed_expression_accessors_report_item_type_and_kind() {
-        let expression =
-            ListExpr::value(vec![Expr::float(FloatExpr::value(1.5))], ValueType::Float)
-                .into_float()
-                .expect("float list");
-
-        assert_eq!(expression.item(), &FloatListItem);
-        assert_eq!(expression.element_type(), ValueType::Float);
-        assert_eq!(
-            expression.kind(),
-            &TypedListExprKind::Value(vec![FloatExpr::value(1.5)]),
-        );
-        assert_eq!(
-            ListExpr::Float(expression.clone()).element_type(),
-            ValueType::Float,
-        );
-        assert_eq!(
-            ListExpr::Float(expression.clone()).into_float(),
-            Some(expression)
-        );
-        assert_eq!(
-            ListExpr::value(vec![Expr::bool(BoolExpr::value(false))], ValueType::Bool).into_float(),
-            None,
-        );
-        assert_eq!(
-            ListExpr::value(
-                vec![Expr::string(StringExpr::value("one".into()))],
-                ValueType::String,
-            )
-            .into_int(),
-            None,
-        );
-        assert_eq!(
-            ListExpr::value(vec![Expr::int(IntExpr::value(1.into()))], ValueType::Int)
-                .into_string(),
-            None,
-        );
-        assert_eq!(
-            ListExpr::value(vec![Expr::int(IntExpr::value(1.into()))], ValueType::Int).into_bool(),
-            None,
-        );
-        assert_eq!(
-            ListExpr::value(vec![Expr::int(IntExpr::value(1.into()))], ValueType::Int).into_nil(),
-            None,
-        );
-        assert_eq!(
-            ListExpr::value(vec![Expr::int(IntExpr::value(1.into()))], ValueType::Int).into_tuple(),
-            None,
-        );
-        assert_eq!(
-            ListExpr::value(vec![Expr::int(IntExpr::value(1.into()))], ValueType::Int).into_list(),
-            None,
-        );
-        assert_eq!(
-            ListExpr::value(vec![Expr::int(IntExpr::value(1.into()))], ValueType::Int)
-                .into_function(),
-            None,
-        );
-    }
-
-    #[test]
     fn facade_dispatch_preserves_every_item_type() {
         let function_type = FunctionType::new(Vec::new(), ValueType::Int);
         let item_types = vec![
@@ -2104,129 +1104,6 @@ mod tests {
     }
 
     #[test]
-    fn bool_case_dispatch_preserves_every_item_type() {
-        let function_type = FunctionType::new(Vec::new(), ValueType::Int);
-
-        assert_eq!(
-            ListExpr::bool_case(
-                BoolExpr::value(true),
-                BoolListCaseBranches::String {
-                    true_: ListExpr::value(Vec::new(), ValueType::String)
-                        .into_string()
-                        .expect("string list"),
-                    false_: ListExpr::value(Vec::new(), ValueType::String)
-                        .into_string()
-                        .expect("string list"),
-                },
-            )
-            .element_type(),
-            ValueType::String,
-        );
-        assert_eq!(
-            ListExpr::bool_case(
-                BoolExpr::value(true),
-                BoolListCaseBranches::Float {
-                    true_: ListExpr::value(Vec::new(), ValueType::Float)
-                        .into_float()
-                        .expect("float list"),
-                    false_: ListExpr::value(Vec::new(), ValueType::Float)
-                        .into_float()
-                        .expect("float list"),
-                },
-            )
-            .element_type(),
-            ValueType::Float,
-        );
-        assert_eq!(
-            ListExpr::bool_case(
-                BoolExpr::value(true),
-                BoolListCaseBranches::Bool {
-                    true_: ListExpr::value(Vec::new(), ValueType::Bool)
-                        .into_bool()
-                        .expect("bool list"),
-                    false_: ListExpr::value(Vec::new(), ValueType::Bool)
-                        .into_bool()
-                        .expect("bool list"),
-                },
-            )
-            .element_type(),
-            ValueType::Bool,
-        );
-        assert_eq!(
-            ListExpr::bool_case(
-                BoolExpr::value(true),
-                BoolListCaseBranches::Nil {
-                    true_: ListExpr::value(Vec::new(), ValueType::Nil)
-                        .into_nil()
-                        .expect("nil list"),
-                    false_: ListExpr::value(Vec::new(), ValueType::Nil)
-                        .into_nil()
-                        .expect("nil list"),
-                },
-            )
-            .element_type(),
-            ValueType::Nil,
-        );
-        assert_eq!(
-            ListExpr::bool_case(
-                BoolExpr::value(true),
-                BoolListCaseBranches::Tuple {
-                    true_: ListExpr::value(Vec::new(), ValueType::Tuple(vec![ValueType::Int]))
-                        .into_tuple()
-                        .expect("tuple list"),
-                    false_: ListExpr::value(Vec::new(), ValueType::Tuple(vec![ValueType::Int]))
-                        .into_tuple()
-                        .expect("tuple list"),
-                },
-            )
-            .element_type(),
-            ValueType::Tuple(vec![ValueType::Int]),
-        );
-        assert_eq!(
-            ListExpr::bool_case(
-                BoolExpr::value(true),
-                BoolListCaseBranches::List {
-                    true_: ListExpr::value(
-                        Vec::new(),
-                        ValueType::List(Box::new(ValueType::String)),
-                    )
-                    .into_list()
-                    .expect("list list"),
-                    false_: ListExpr::value(
-                        Vec::new(),
-                        ValueType::List(Box::new(ValueType::String)),
-                    )
-                    .into_list()
-                    .expect("list list"),
-                },
-            )
-            .element_type(),
-            ValueType::List(Box::new(ValueType::String)),
-        );
-        assert_eq!(
-            ListExpr::bool_case(
-                BoolExpr::value(true),
-                BoolListCaseBranches::Function {
-                    true_: ListExpr::value(
-                        Vec::new(),
-                        ValueType::Function(Box::new(function_type.clone())),
-                    )
-                    .into_function()
-                    .expect("function list"),
-                    false_: ListExpr::value(
-                        Vec::new(),
-                        ValueType::Function(Box::new(function_type.clone())),
-                    )
-                    .into_function()
-                    .expect("function list"),
-                },
-            )
-            .element_type(),
-            ValueType::Function(Box::new(function_type)),
-        );
-    }
-
-    #[test]
     fn nested_list_index_dispatch_preserves_every_item_type() {
         let function_type = FunctionType::new(Vec::new(), ValueType::Int);
         let item_types = vec![
@@ -2313,117 +1190,6 @@ mod tests {
             )
             .element_type(),
             ValueType::Function(Box::new(function_type)),
-        );
-    }
-
-    #[test]
-    fn metadata_helpers_preserve_tuple_list_and_function_item_types() {
-        let tuple_item = TupleListItem {
-            item_type: vec![ValueType::Int],
-        };
-        let list_item = ListListItem {
-            item_type: Box::new(ValueType::String),
-        };
-        let function_type = FunctionType::new(Vec::new(), ValueType::Bool);
-        let function_item = FunctionListItem {
-            item_type: function_type.clone(),
-        };
-
-        assert_eq!(
-            into_nil_clauses(vec![(
-                BigInt::from(1),
-                ListExpr::value(vec![Expr::nil(NilExpr::value())], ValueType::Nil),
-            )]),
-            vec![(
-                BigInt::from(1),
-                ListExpr::value(vec![Expr::nil(NilExpr::value())], ValueType::Nil)
-                    .into_nil()
-                    .expect("nil list"),
-            )],
-        );
-        assert_eq!(
-            into_tuple_clauses(
-                vec![(
-                    "tuple",
-                    ListExpr::value(
-                        vec![Expr::tuple(TupleExpr::value(
-                            vec![Expr::int(IntExpr::value(1.into()))],
-                            vec![ValueType::Int],
-                        ))],
-                        ValueType::Tuple(vec![ValueType::Int]),
-                    ),
-                )],
-                &tuple_item,
-            ),
-            vec![(
-                "tuple",
-                ListExpr::value(
-                    vec![Expr::tuple(TupleExpr::value(
-                        vec![Expr::int(IntExpr::value(1.into()))],
-                        vec![ValueType::Int],
-                    ))],
-                    ValueType::Tuple(vec![ValueType::Int]),
-                )
-                .into_tuple()
-                .expect("tuple list"),
-            )],
-        );
-        assert_eq!(
-            into_list_clauses(
-                vec![(
-                    "list",
-                    ListExpr::value(
-                        vec![Expr::list(ListExpr::value(
-                            vec![Expr::string(StringExpr::value("one".into()))],
-                            ValueType::String,
-                        ))],
-                        ValueType::List(Box::new(ValueType::String)),
-                    ),
-                )],
-                &list_item,
-            ),
-            vec![(
-                "list",
-                ListExpr::value(
-                    vec![Expr::list(ListExpr::value(
-                        vec![Expr::string(StringExpr::value("one".into()))],
-                        ValueType::String,
-                    ))],
-                    ValueType::List(Box::new(ValueType::String)),
-                )
-                .into_list()
-                .expect("list list"),
-            )],
-        );
-        assert_eq!(
-            into_function_clauses(
-                vec![(
-                    "function",
-                    ListExpr::value(
-                        vec![Expr::function(FunctionExpr::value(FunctionValue::new(
-                            RuntimeFunctionId::Bool(crate::plan::BoolFunctionId(0)),
-                            Vec::new(),
-                        )))],
-                        ValueType::Function(Box::new(FunctionType::new(
-                            Vec::new(),
-                            ValueType::Bool
-                        ))),
-                    ),
-                )],
-                &function_item,
-            ),
-            vec![(
-                "function",
-                ListExpr::value(
-                    vec![Expr::function(FunctionExpr::value(FunctionValue::new(
-                        RuntimeFunctionId::Bool(crate::plan::BoolFunctionId(0)),
-                        Vec::new(),
-                    )))],
-                    ValueType::Function(Box::new(FunctionType::new(Vec::new(), ValueType::Bool))),
-                )
-                .into_function()
-                .expect("function list"),
-            )],
         );
     }
 }
